+++ src/java/org/apache/cassandra/service/StorageService.java
@@ -47,9 +47,9 @@ import ch.qos.logback.classic.LoggerContext;
 import ch.qos.logback.classic.jmx.JMXConfiguratorMBean;
 import ch.qos.logback.classic.spi.ILoggingEvent;
 import ch.qos.logback.core.Appender;
-import com.palantir.cassandra.cvam.CrossVpcIpMappingAckVerbHandler;
-import com.palantir.cassandra.cvam.CrossVpcIpMappingHandshaker;
-import com.palantir.cassandra.cvam.CrossVpcIpMappingSynVerbHandler;
+import com.palantir.cassandra.cvim.CrossVpcIpMappingAckVerbHandler;
+import com.palantir.cassandra.cvim.CrossVpcIpMappingHandshaker;
+import com.palantir.cassandra.cvim.CrossVpcIpMappingSynVerbHandler;
 import org.apache.cassandra.auth.AuthKeyspace;
 import org.apache.cassandra.auth.AuthMigrationListener;
 import org.apache.cassandra.concurrent.ScheduledExecutors;
