+++ test/unit/com/palantir/cassandra/cvim/SerializationsTest.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package com.palantir.cassandra.cvam;
+package com.palantir.cassandra.cvim;
 
 import java.io.DataInputStream;
 import java.io.IOException;
